+++ server/service/copse.js
@@ -1,4 +1,9 @@
 return {
+	payload(record, output){
+		const payload = this.crdt.prototype.stringify(record)
+		Object.assign(output, {payload})
+		return this.next()
+	},
 	async router(method, params){
 		const idx = params.id ? '/id' : ''
 		const name = `${method}/chat${idx}`
